requirements.txt
pydantic==2.10.4
pydantic_ai==0.0.15
python-dotenv==1.0.1
termcolor==2.5.0
